+++ packages/esp_board_manager/devices/dev_audio_codec/dev_audio_codec.c
@@ -347,13 +347,8 @@ int dev_audio_codec_deinit(void *device_handle)
     }
     audio_codec_delete_data_if(codec_handles->data_if);
 
-    const char *name = NULL;
-    const esp_board_device_handle_t *device_handle_struct = esp_board_device_find_by_handle(device_handle);
-    if (device_handle_struct) {
-        name = device_handle_struct->name;
-    }
     dev_audio_codec_config_t *cfg = NULL;
-    esp_board_device_get_config(name, (void **)&cfg);
+    esp_board_device_get_config_by_handle(device_handle, (void **)&cfg);
     if (cfg) {
         esp_board_periph_unref_handle(cfg->i2s_cfg.name);
         esp_board_periph_unref_handle(cfg->i2c_cfg.name);