+++ packages/esp_board_manager/devices/dev_lcd_touch_i2c/dev_lcd_touch_i2c.c
@@ -111,13 +111,8 @@ int dev_lcd_touch_i2c_deinit(void *device_handle)
         touch_handles->io_handle = NULL;
     }
 
-    const char *name = NULL;
-    const esp_board_device_handle_t *device_handle_struct = esp_board_device_find_by_handle(device_handle);
-    if (device_handle_struct) {
-        name = device_handle_struct->name;
-    }
     dev_lcd_touch_i2c_config_t *cfg = NULL;
-    esp_board_device_get_config(name, (void **)&cfg);
+    esp_board_device_get_config_by_handle(device_handle, (void **)&cfg);
     if (cfg) {
         esp_board_periph_unref_handle(cfg->i2c_name);
     }